// src/lib/types/factbook.ts
export interface FactbookDetails {
    id: string;
    title: string;
    category: string;
    subcategory: string;
    author: string;
    created: number;
    edited: number;
    views: number;
    score: number;
}
export interface Dispatch {
    id: string;
    title: string;
    category: string;
    subcategory: string;
    author: string;
    created: number;
    edited: number;
    views: number;
    score: number;
    text: string;
}